+++ frontend/src/domains/user/OnBoarding/OnBoarding.tsx
@@ -1,6 +1,6 @@
-import { getOrganizationName } from '@/apis/organization.api';
 import BasicButton from '@/components/BasicButton/BasicButton';
 import SkipIcon from '@/components/icons/SkipIcon';
+import useOrganizationName from '@/domains/hooks/useOrganizationName';
 import {
   container,
   place,
@@ -12,7 +12,6 @@ import {
   title,
 } from '@/domains/user/OnBoarding/OnBoarding.styles';
 import { useAppTheme } from '@/hooks/useAppTheme';
-import { useEffect, useState } from 'react';
 
 interface OnBoardingProps {
   moveNextStep: () => void;
@@ -21,15 +20,7 @@ interface OnBoardingProps {
 export default function OnBoarding({ moveNextStep }: OnBoardingProps) {
   const theme = useAppTheme();
 
-  const [placeName, setPlaceName] = useState('피드줍줍');
-
-  useEffect(() => {
-    async function getOrganization() {
-      const response = await getOrganizationName({ organizationId: 1 });
-      setPlaceName(response!.data.organizationName);
-    }
-    getOrganization();
-  }, []);
+  const { placeName } = useOrganizationName();
 
   return (
     <section css={container}>
